+++ PSServicePrincipal/functions/Get-SpnByAppID.ps1
@@ -10,10 +10,20 @@
         .PARAMETER ApplicationID
             This parameter is the application id of the objects you are retrieving.
 
+        .PARAMETER EnableException
+            This parameter disables user-friendly warnings and enables the throwing of exceptions.
+            This is less user friendly, but allows catching exceptions in calling scripts.
+
         .EXAMPLE
             PS c:\> Get-SpnByAppID -ApplicationID 34a23ad2-dac4-4a41-bc3b-d12ddf90230e
 
             This will retrieve a Service Principal by application id from the Azure active directory.
+
+        .EXAMPLE
+            PS c:\> Get-SpnByAppID -ApplicationID 34a23ad2-dac4-4a41-bc3b-d12ddf90230e -EnableException
+
+            This example gets a service principal in AAD, after prompting for user preferences.
+            If this execution fails for whatever reason (connection, bad input, ...) it will throw a terminating exception, rather than writing the default warnings.
     #>
 
     [OutputType('System.String')]
@@ -22,7 +32,10 @@
         [parameter(Mandatory = 'True', Position = '0', HelpMessage = "ApplicationID used to retrieve an application")]
         [ValidateNotNullOrEmpty()]
         [string]
-        $ApplicationID
+        $ApplicationID,
+
+        [switch]
+        $EnableException
     )
 
     try