+++ PSServicePrincipal/internal/functions/Connect-ToCloudTenant.ps1
@@ -17,7 +17,7 @@
         .EXAMPLE
             PS c:\> Connect-ToCloudTenant -Reconnect -EnableException
 
-            This will be used to make a connection to an Azure tenant or reconnect to another specified tenant.
+            This example makes a connection to an Azure tenant or reconnect to another specified tenant.
             If this execution fails for whatever reason (connection, bad input, ...) it will throw a terminating exception, rather than writing the default warnings.
     #>